Move the checkbox controls out of the sidebar
sidebar.tsx had become the de facto module for CheckboxSetting and
CheckboxSettingCustom: nine of its consumers -- the mapper's four
editors, the quiz, so-you're-an-urbanist, the histogram, the related
buttons, the statistic tree -- import nothing else from it, and none of
them are anywhere near a sidebar. Only template.tsx actually wants the
Sidebar.

Pure move, no behavior change; the components are unchanged apart from
CheckboxSettingJustBox and its props interface becoming exported, since
they're part of what the new module offers.
