+++ react/src/syau/syau-panel.tsx
@@ -3,9 +3,9 @@ import React, { ReactNode, useContext, useRef } from 'react'
 import '../common.css'
 
 import { buttonStyle, GenericShareButton } from '../components/GenericShareButton'
+import { CheckboxSetting } from '../components/checkbox-setting'
 import { CountsByUT } from '../components/countsByArticleType'
 import { createScreenshot } from '../components/screenshot'
-import { CheckboxSetting } from '../components/sidebar'
 import { Navigator } from '../navigation/Navigator'
 import { useColors, useJuxtastatColors } from '../page_template/colors'
 import { useSetting } from '../page_template/settings'
